Fixed apparent redundancy

+++ controllers/itemController.js
@@ -1,4 +1,5 @@
 const asyncHandler = require('express-async-handler');
+
 const Item = require('../models/Item');
 
 
@@ -6,15 +7,6 @@ const getItems = asyncHandler(async(req, res) => {
     
 
     const items = await Item.find().sort({updatedAt: -1});
-
-    if(!items){
-
-        return res.status(404).json({message: "No items were found"});
-    }
-
-    if(items && items.length == 0){
-        return res.status(200).json({message: "There are no items"})
-    }
 
     res.status(200).json(items);
 
@@ -24,14 +16,30 @@ const getItems = asyncHandler(async(req, res) => {
 })
 
 
+const getItem = asyncHandler(async(req, res) => {
+    
+    const {id} = req.params.id;
+    const item = await Item.findById(id);
+
+    res.status(200).json(item);
+
+
+
+    
+})
+
+
+
 const createItem = asyncHandler(async(req, res) => {
 
     const user = req.user;
 
 
+    
     if((user.roles.indexOf("admin") === -1) && (user.roles.indexOf("vendor") === -1)){
         return res.status(403).json({message: "User is neither vendor nor admin"});
-    }
+        
+    } 
 
 
 
@@ -42,7 +50,7 @@ const createItem = asyncHandler(async(req, res) => {
     const item = await Item.create({name, brand, model, category, price});
 
     if(!item){
-        return res.status(404).json("Item could not be create. Something went wrong");
+        return res.status(404).json("Item could not be created. Something went wrong");
     }
     
 
@@ -56,34 +64,34 @@ const updateItem = asyncHandler(async(req, res) => {
 
 
     const user = req.user;
+    
+
+    
     if((user.roles.indexOf("admin") === -1) && (user.roles.indexOf("vendor") === -1)){
         return res.status(403).json({message: "User is neither vendor nor admin"});
-    }
-
+        
+    } 
    const {id} = req.params;
 
    const {name, brand, model, category, price} = req.body;
 
-   const item = await Item.findById(id).exec();
+   const item = await Item.findById(id);
 
    if(!item){
 
     return res.status(404).json({message: "Item does not exist"});
    }
-   
-   item.name = name ? name : item.name;
-   item.brand = brand ? brand : item.brand;
-   item.model = model ? model : model.brand;
-   item.category = category ? category : item.category;
-   item.price = price ? price : item.price;
-   
+       
 
-   const updateSuccessful = await item.save();
+   const updateSuccessful = await Item.findByIdAndUpdate(id, {name, brand, model, category, price}, {new: true});
+
 
    if(!updateSuccessful){
 
     return res.status(404).json({message: "Update was not successful"});
    }
+   
+
 
    res.status(200).json({message: "Update was successful", updateSuccessful});
 
@@ -95,6 +103,12 @@ const updateItem = asyncHandler(async(req, res) => {
 const deleteItem = asyncHandler(async(req, res) => {
   
     const {id} = req.params;
+    const user = req.user;
+
+    if((user.roles.indexOf("admin") === -1) && (user.roles.indexOf("vendor") === -1)){
+        return res.status(403).json({message: "User is neither vendor nor admin"});
+        
+    } 
 
 
     const item = await Item.findOneAndDelete({id});
@@ -116,6 +130,7 @@ const deleteItem = asyncHandler(async(req, res) => {
 module.exports = {
 
     getItems,
+    getItem,
     createItem,
     updateItem,
     deleteItem
